Rename eos_data_distribution module to eosdatadistribution
Underscores are discouraged by PEP 8:
https://www.python.org/dev/peps/pep-0008/#package-and-module-names

They also break pybuild building a Debian package: the Python side of
things wants the package names to be python-eos_data_distribution; but
Debian does not allow underscores in package names.


https://phabricator.endlessm.com/T14639
