+++ eosdatadistribution/store/ostree_store.py
@@ -24,9 +24,9 @@ import gi
 from gi.repository import GObject
 from gi.repository import GLib
 
-from eos_data_distribution import SimpleStore
-from eos_data_distribution.names import SUBSCRIPTIONS_INSTALLED
-from eos_data_distribution.subscription import Producer as SubscriptionProducer
+from eosdatadistribution import SimpleStore
+from eosdatadistribution.names import SUBSCRIPTIONS_INSTALLED
+from eosdatadistribution.subscription import Producer as SubscriptionProducer
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
